backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]==3.3.0
python-multipart
